requirements.txt
Click
PyGObject
copypaste
xerox
pyyaml